+++ babel.config.js
@@ -2,10 +2,20 @@ const path = require('path');
 const { babelAliases } = require('./config/aliases');
 
 const config = {
-    presets: ['@babel/preset-typescript'],
+    presets: [
+        '@babel/preset-typescript',
+        [
+            '@babel/env',
+            {
+                useBuiltIns: false,
+                loose: true,
+                modules: process.env.BABEL_ENV === 'es' ? false : 'commonjs',
+                bugfixes: true,
+            },
+        ],
+    ],
     plugins: [
         '@babel/proposal-class-properties',
-        '@babel/proposal-object-rest-spread',
         '@babel/plugin-proposal-nullish-coalescing-operator',
         [
             'babel-plugin-custom-import-path-transform',
@@ -13,7 +23,6 @@ const config = {
                 transformImportPath: path.resolve(__dirname, 'scripts/transformImportPath.js'),
             },
         ],
-        '@babel/plugin-transform-runtime',
         [
             'babel-plugin-module-resolver',
             {
@@ -21,24 +30,17 @@ const config = {
                 root: ['./src'],
             },
         ],
+        '@babel/plugin-proposal-object-rest-spread',
+        [
+            '@babel/plugin-transform-runtime',
+            {
+                useESModules: false,
+                regenerator: false,
+                helpers: true,
+            },
+        ],
     ],
     ignore: process.env.BABEL_ENV === 'test' ? [] : ['**/__tests__/', '**/*.test.ts'],
 };
-
-if (process.env.BABEL_ENV === 'es') {
-    config.presets.push([
-        '@babel/modules',
-        {
-            loose: true,
-        },
-    ]);
-} else {
-    config.presets.push([
-        '@babel/env',
-        {
-            loose: true,
-        },
-    ]);
-}
 
 module.exports = config;