+++ src/reducers/api/infiniteList/factoryReducer.ts
@@ -1,7 +1,7 @@
 import { config, InfiniteReducerState, Action, undefinedActionTypesWarning } from '../../../config';
 import { CustomParams } from '../../types';
 
-import makeBasicApiReducer from '../basic';
+import makeBasicApiReducer from '../basic/factoryReducer';
 import {
     actionFilters as basicApiReducerActionFilters,
     actionTypes as basicApiReducerActionTypes,
@@ -32,6 +32,36 @@ const getParams = (customParams: CustomParams = {}) => {
     };
 };
 
+/**
+ * Docs: https://github.com/AckeeCZ/redux-utils/blob/master/docs/reducers/infiniteListApiReducer.md
+ * 
+ * @param {object} params 
+    * @param {object} params.actionTypes 
+        * @param {(string | undefined)} params.actionTypes.REQUEST
+        * @param {(string | undefined)} params.actionTypes.CANCEL
+        * @param {(string | undefined)} params.actionTypes.SUCCESS
+        * @param {(string | undefined)} params.actionTypes.FAILURE
+        * @param {(string | undefined)} params.actionTypes.RESET
+        * @param {(string | undefined)} params.actionTypes.UPDATE
+
+    * @param {object} [params.initialState]
+        * @param {boolean} params.initialState.inProgress
+        * @param {any} params.initialState.error
+        * @param {boolean} params.initialState.success
+        * @param {boolean} params.initialState.cancelled
+        * @param {(number | null)} params.initialState.lastSuccessAt
+        * @param {boolean} params.initialState.hasMore
+        * @param {number} params.initialState.payloadSize
+        * @param {number} params.initialState.totalOffset
+        
+     * @param {object} [params.selectors]
+        * @param {(action: object) => number} [params.selectors.currentCount]
+        
+    * @param {object} [params.actionFilters]
+        * @param {(action: object) => boolean} [params.actionFilters.update]
+
+ * @returns {(state: object, action: object) => object}
+ */
 export default function makeInfiniteListApiReducer(customParams: CustomParams) {
     const { actionTypes: types, initialState, selectors, options, actionFilters }: CustomParams = getParams(
         customParams,
@@ -58,12 +88,14 @@ export default function makeInfiniteListApiReducer(customParams: CustomParams) {
 
             case types.SUCCESS: {
                 const currentCount = selectors.currentCount(action);
+                const totalCount = selectors.totalCount(action);
 
                 return {
                     ...state,
                     ...basicApiReducer(state, action),
                     totalOffset: state.totalOffset + currentCount,
                     hasMore: currentCount >= state.payloadSize,
+                    totalCount,
                 };
             }